hs9-flask-api-raamatud/requirements.txt
azure-core==1.30.1
azure-storage-blob==12.20.0b1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Flask==3.0.3
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
pycparser==2.22
requests==2.31.0
six==1.16.0
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.2
flask-cors
